Documents/GitHub/InsightPulseAI_SKR/tools/js/pulser-cli/requirements.txt
# Core dependencies
pyyaml>=6.0
requests>=2.28.1
python-dotenv>=0.21.0

# Tool integration
jsonschema>=4.17.3
rich>=13.3.1